Clean up compile warning on iofunc.c
(cherry picked from commit 7455fa1b918527b1ae8395883cc57022adb784f0)

+++ src/iofunc.c
@@ -665,7 +665,7 @@ int HAMLIB_API read_string(hamlib_port_t *p,
     struct timeval tv, tv_timeout, start_time, end_time, elapsed_time;
     int total_count = 0;
 
-    rig_debug(RIG_DEBUG_TRACE, "%s called, rxmax=%ld\n", __func__, rxmax);
+    rig_debug(RIG_DEBUG_TRACE, "%s called, rxmax=%d\n", __func__, (int)rxmax);
 
     if (!p || !rxbuffer)
     {
